Fix multiple malloc and free bugs
This was a hard bug to fix as you can see from the extra debugging tools
I wrote to fix it:

* I wrote `i64tohex` for a more readable int-to-str function
* The standard library now contains a `verify_memory` function which
  walks through the memory after every free and malloc (but only when a
  toggle is set).
* Another toggle is added to dump the metadata of all the dynamically
  allocated memory.

Very handy, will probably be useful later.

One of the bugs was in the free function when free blocks were merged,
another bug was during the splitting of a free block where the size of a
block shouldn't be set when a split couldn't occur (the original block
is bigger than the new block, but not big enough for two).