Add traceback flowchart
In a language as finicky as Ilo, we could really use all the help when
troubleshooting, and creating good error messages is hard to do.

When something happens, the first thing I usually do is try to figure
out what the traceback is.. by hand.. This needs to be improved. Hence
this flowchart.

Cheers! 🍻