Push traceback string and main flag with each function call
In order to support tracebacks, we want to push a message to the stack
before calling a function, so that we know where the call originated
from. We also push a boolean which indicates if this is the root
function call to main, so that we know where to stop the traceback.