Add function to check if list contains string
There will be a lot of times an Ilo program uses lists, and a lot of
those lists will contain strings. My guess is that a lot of these lists
of strings might be checked if they contain a certain string, so this
commit adds a function to do just that to the standard library.