Change token index into local variable
The token index of the parser is a global variable. However, this is not
what we want as the token index should not be shared between parsed
source files (in other words, imported files). This commit fixes that by
changing this index from a global variable to a local one.