Parse if/elif/else statements!
I've been saving this one until the end (well.. almost), but it was a
fun one. Encountered my first overflow bug due to an incorrect malloc
(see previous commit) for which I extensively used verify_memory. It
turns out to be a great tool to have.

Glad it works now, almost done with parsing!