Fix off-by-one error in concat
It's a miracle this hasn't caused any memory bugs before and it's an
even bigger miracle we're finding this just before we finish the parser.

The null-terminator byte of a concatenated string was placed one offset
too far. This hasn't been noticed before, because concat mallocs
apparently haven't used reused memory before.