Replace redundant constants struct by dictionary
Now that we've got hashmaps, let's use them in our self-hosted compiler!

A nice side-effect of this commit is that the `while_skip` test is
failing due to a dynamic memory allocation bug that I've been trying to
fix for the last couple of days. Now we have a small test to reproduce
the issue, without the bulky standard library in our way!