Add textbuffer class to standard library
We're currently malloc'ing and free'ing everytime we append text to a
buffer in the code generator. Almost every byte that's outputted is
written as many times as we have opcodes, so our code generator takes
O(n^2) time! Terrible!

A textbuffer like this improves it a lot; it only takes O(n) to append
text!