Remove most malloc's and concat's from code generator
Concatenating strings and malloc'ing takes us between O(n) and O(n^2)
time, but we can make it much better by allocating buffers on forehand
and filling them in about O(1) time (so the whole IR is converted in
O(n)). MUCH better!

There are still some concat's we need to work around, but this is the
first time we've succesfully compiled the standard library in an almost
acceptable amount of time!! WOOHOOOOO!