Fix lexer bug
It's very interesting that this bug hasn't been noticed before as this
should happen often, but better late than never.

The `is_word` function caused a segmentation fault when the length of
string `q` is more than the length of string `s`. This commit adds that
case to prevent the segmentation fault.