Do not define duplicate strings
The compiled compiler assembly code contains a lot of duplicate string
definitions, which is a waste of time and space. We prevent duplicates
by adding a dictionary with the string as key and its label as value.

While this does reduce the number of lines of the assembly code, it's
even better when the function call string is replaced with a function
string and an integer for the line number.