Implement a profiler
We want to optimize our compiler and need to know which functions take
up the largest amount of time.

For this, I built a simple profiler. It samples the call stack at 1kHz
using linux timer signals and saves a list of names of the calling
functions (so not the called functions, we might want to add that data
to the function call stub).

Of course, this is the minimal amount of information as keeping track of
the full call stack would be more informative, but this is simpler and
might be enough for our current goal of optimizing the compiler.

Also, I've added a flag to the self-hosted compiler to enable this
profiler.