Add gcd and ratios to the standard library
We'll need to be able to express and calculate non-integer numbers if we
ever want to create a pathtracer. While floating points are the right
way to go, they are relatively hard to implement. That's a problem,
because I'm lazy.

Fractions are way easier; both the data structure and the arithmetic
functions are almost trivial to implement. Sounds good!